StructDiff

Documentation for StructDiff.

StructDiff.AbstractDiff

StructDiff.ArrayDiff

StructDiff.AtomicDiff

StructDiff.BitsDiff

StructDiff.DictDiff

StructDiff.DiffCollection

StructDiff.FieldsDiff

StructDiff.NamedDiff

StructDiff.SizeDiff

StructDiff.TypeDiff

StructDiff.compare

StructDiff.nodiff

StructDiff.@test_diff

StructDiff.AbstractDiff — Type

General abstract representation of a difference between two objects. Any AbstractDiff type should implement the following:

printdiff(::IO, ::AbstractDiff)

nodiff(::AbstractDiff)::Bool whether there are differences to consider.

Optionally you can implement:

xcolor::Symbol color of the left object (default :blue).

ycolor::Symbol color of the right object (default :yellow).

StructDiff.ArrayDiff — Type

Collection of array element differences.

StructDiff.AtomicDiff — Type

An AtomicDiff is a leaf in the diff tree, i.e. it does not have any children differences. AtomicDiff subtypes should implement:

Base.first(::AtomicDiff): the left element

Base.last(::AtomicDiff): the right element

Subtypes can also implement the diff_prefix(::IO, ::AtomicDiff) to print some additional prefix.

StructDiff.BitsDiff — Type

Basic representation of different objects, also used as a fallback.

StructDiff.DictDiff — Type

Collection of comparison of (key, value) pairs.

StructDiff.DiffCollection — Type

As opposed to AtomicDiff, DiffCollection contains multiple diff objects.

You should implement vals(::DiffCollection) to return the unfiltered collection of diff elements. Note that it defaults to getproperty(::DiffCollection, :diffs).

StructDiff.FieldsDiff — Type

Collection of fields differences, usually comes from comparing struct.

StructDiff.NamedDiff — Type

Difference with a specific name indicator. The key can be of all types.

StructDiff.SizeDiff — Type

Representation of different sizes.

StructDiff.TypeDiff — Type

Representation of different types.

StructDiff.compare — Method

compare(x, y) -> AbstractDiff

Main functionality of the package. Takes two argument and build a comparison tree that recursively checks different aspects of the compared objects. The obtained object can be checked for equality with nodiff.

StructDiff.nodiff — Method

Indicates if the given AbstractDiff object contains actual differences.

StructDiff.@test_diff — Macro

Equivalent to the @test macro, that will print the difference tree if @test_diff is called on an equality call.